refactor(stream): extract StreamView helpers into focused modules
Move formatters, indicators, title bar, and empty-state helpers into
components/stream/ without changing StreamView behavior.
