+++ opennow-stable/src/renderer/src/components/stream/streamFormatters.ts
@@ -1,0 +1,27 @@
+import { formatElapsed } from "../../utils/timeFormat";
+
+export function formatFileSize(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+}
+
+export function formatWarningSeconds(value: number | undefined): string | null {
+  if (value === undefined || !Number.isFinite(value) || value < 0) {
+    return null;
+  }
+  const total = Math.floor(value);
+  const minutes = Math.floor(total / 60);
+  const seconds = total % 60;
+  if (minutes > 0) {
+    return `${minutes}m ${seconds.toString().padStart(2, "0")}s`;
+  }
+  return `${seconds}s`;
+}
+
+export function formatSessionTimeRemaining(value: number | null): string | null {
+  if (value === null || !Number.isFinite(value) || value < 0) {
+    return null;
+  }
+  return formatElapsed(value);
+}
